Refactor: 주간완료 할일 계산함수 분리
